Some coding style changes, made one method const.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@1242 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CNullDriver.cpp
@@ -1500,12 +1500,13 @@ E_DRIVER_TYPE CNullDriver::getDriverType() const
 void CNullDriver::deleteMaterialRenders()
 {
 	// delete material renderers
-	for (int i=0; i<(int)MaterialRenderers.size(); ++i)
+	for (u32 i=0; i<MaterialRenderers.size(); ++i)
 		if (MaterialRenderers[i].Renderer)
 			MaterialRenderers[i].Renderer->drop();
 
 	MaterialRenderers.clear();
 }
+
 
 //! Returns pointer to material renderer or null
 IMaterialRenderer* CNullDriver::getMaterialRenderer(u32 idx)
@@ -1515,7 +1516,6 @@ IMaterialRenderer* CNullDriver::getMaterialRenderer(u32 idx)
 	else
 		return 0;
 }
-
 
 
 //! Returns amount of currently available material renderers.
